// src/typing/Interfaces/IReactHookDataForm.ts
export interface IDataForm{
  Nome: string,
  Cpf: string
  Whatsapp: string,
  Estado: string,
  Cidade: string,
  Endereco: string,
  Complemento: string,
  Numero: string,
  Cep: string,
  Bairro: string
  email: string
  CardNumber: string
  CardExpire: string
  CardName: string
  CardCVC: string
  parcelas: number
}